[JIT] Add recursive scripting for class type module attributes
**Summary**
This commit modifies type inference (used by the module scripting code)
so that it tries to script the type of any class instances that it
encounters. This enables recursive, automatic scripting of class type
module attributes. If the type cannot be successfully compiled, the
module attribute that triggered its compilation is added to the concrete
type of the module as a failed attribute. Any scripted code that uses it
will fail to compile with an error message that explains that
TorchScript tried to covert the attribute of the problematic class type
and failed.

**Test Plan**
This commit adds test cases for this to `TestClassType`, one for
successful recursive scripting and another for unsuccessful recursive
scripting.

ghstack-source-id: b66490018acc285937010d8a14b0d9699f961d8a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45234